rename back the znet.h to znet.hpp
