+++ cpp/include/znn/jit/Makefile
@@ -31,7 +31,7 @@ STANDARD_FLAG	=	-std=c++14
 STD_FLAG		=
 
 CXX             =   $(AT)icpc $(STANDARD_FLAG)                                           
-OPTIMIZATION_FLAGS  =   -DNDEBUG -O3 -ipo -fp-model fast=2 -mkl=sequential -fno-rtti -fno-exceptions 
+OPTIMIZATION_FLAGS  =   -DDEBUG -O3 -ipo -fp-model fast=2 -mkl=sequential -fno-rtti -fno-exceptions 
 CS_FLAGS        =   -mkl=sequential                                                       
 CS_LD_FLAGS     =   -mkl=sequential  -static-intel 
 
@@ -52,7 +52,7 @@ COMMON_FLAGS	=	$(INLINE_DEPFLAGS) $(CXXINCLUDES) $(CXXWARN) $(CXXERR)
 
 DBG_FLAGS		=	$(COMMON_FLAGS) $(CS_FLAGS)
 
-LD_OPTIMIZATION_FLAGS	=	-DNDEBUG -O3
+LD_OPTIMIZATION_FLAGS	=	-DDEBUG -O3
 COMMON_LDFLAGS		=	$(HBW_LD_FLAG)
 
 OPT_FLAGS		=	$(COMMON_FLAGS) $(OPTIMIZATION_FLAGS) $(CS_FLAGS)
