+++ cpp/main.cpp
@@ -1,4 +1,4 @@
-#include "znet.h"
+#include "znet.hpp"
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
 #include <pybind11/stl.h>
